transaction and monitoring services with fraud rules implemented

+++ src/main/java/com/hackathon/finservice/dto/AccountDashboardResponse.java
@@ -1,8 +1,10 @@
 package com.hackathon.finservice.dto;
 
+import java.math.BigDecimal;
+
 public record AccountDashboardResponse(
         String accountNumber,
-        double balance,
+        BigDecimal balance,
         String accountType
     ) {
 }
